refactor : 재료 관련 리팩토링 (#37) - 재료테이블에서 unit을 재료 중간 테이블로 이동 - amount을 integer에서 double로 수정 및 관련 메서드 & dto 수정 - 프롬프트 레시피 제목 정규화해서 수정

+++ src/main/java/com/shortkki/api/recipe/dto/IngredientRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 public record IngredientRequest(
         @NotBlank(message = "재료명은 필수입니다") String name,
         @NotNull(message = "단위는 필수입니다") String unit,
-        Integer amount
+        Double amount
 ) {
 
 }
