+++ src/main/java/com/shortkki/api/recipe/dto/IngredientResponse.java
@@ -3,7 +3,7 @@ package com.shortkki.api.recipe.dto;
 public record IngredientResponse(
         String name,
         String unit,
-        Integer amount
+        Double amount
 ) {
 
 }
